+++ src/api/plaid/accounts/get/index.ts
@@ -92,7 +92,7 @@ router.get('/', async (req: any, res: any, next: any) => {
                 method_used: req.method,
             }
         };
-        console.log('INSIDE CATCH');
+        console.error('INSIDE CATCH');
         finalStatus = 400;
     }
     await res.status(finalStatus);