+++ src/api/plaid/transactions/categories/index.ts
@@ -42,7 +42,7 @@ router.post('/', async (req: any, res: any, next: any) => {
           method_used: req.method,
       }
   };
-  console.log('INSIDE CATCH');
+  console.error('INSIDE CATCH');
   finalStatus = 400;
   }
   await res.status(finalStatus);