+++ src/api/plaid/transactions/get/index.ts
@@ -131,7 +131,7 @@ router.get('/', async (req:any, res:any, next:any) => {
           method_used: req.method,
       }
   };
-  console.log('INSIDE CATCH');
+  console.error('INSIDE CATCH');
   finalStatus = 400;
   };
   await res.status(finalStatus);